Modified ToString Return
I added small comma here and there in the first part, and redid everything over again in the 2nd part.
Not sure how much is broken. I didnt do any testing. Please double check my counting on `string.Format`, i'm not sure if it's really 33 values...

The idea is, if it doesnt have any value, it doesnt need to be there. So i added a null check, and then told it to skip. I added commas before each value. Added a splash of `\n`, for clarity and readability. They arent a lot, and they're not needed. (So they can easily be removed)